VISTA-2014 remove BMO file from server it it is OK but not from production

+++ vista-shared/vista-payment/src/main/java/com/propertyvista/payment/dbp/BmoSftpClient.java
@@ -34,7 +34,9 @@ public class BmoSftpClient {
     }
 
     public void removeFile(String fileName) {
-        SftpClient.removeFile(configuration(), ".", fileName);
+        if (configuration().removeReceivedFileFromSftpHost()) {
+            SftpClient.removeFile(configuration(), ".", fileName);
+        }
     }
 
 }
